Implementado ciclo completo da entidade Torneio (Entity, Repository, Sevice e Controller) e mapeamento com banco

+++ src/main/java/br/com/esports/arenas/repository/TorneioRepository.java
@@ -3,6 +3,6 @@ package br.com.esports.arenas.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import br.com.esports.arenas.model.Torneio;
 
-public interface TorneioRepository extends JpaRepository<Torneio, Long> {
+public interface TorneioRepository extends JpaRepository<Torneio, Integer> {
 
 }
